the end (maybe)

+++ src/fastapi_test_task/services/user_service.py
@@ -3,14 +3,23 @@ from sqlalchemy.orm import Session
 from typing import List
 from repositories.user_repository import UserRepositoryCreate, UserRepositoryUpdate
 
-from tables import User
+from tables import User, Book
 from database import get_session
 
 class UserService:
     def __init__(self, session: Session = Depends(get_session)):
         self.session = session
 
-    def _get(self, user_id: int) -> User:
+    def __has_book(self, user_id: int) -> bool:
+        book = (
+            self.session
+            .query(Book)
+            .filter_by(user_id=user_id)
+            .first()
+        )
+        return True if book else False
+
+    def __get(self, user_id: int) -> User:
         user = (
             self.session
             .query(User)
@@ -30,7 +39,7 @@ class UserService:
         return users
 
     def get(self, user_id: int) -> User:
-        return self._get(user_id)   
+        return self.__get(user_id)   
 
     def create(self, user_data: UserRepositoryCreate) -> User:
         user = User(**user_data.dict())
@@ -38,15 +47,20 @@ class UserService:
         self.session.commit()
         return user
 
-    def update(self, user_id: int, user_data: UserRepositoryUpdate):
-        user = self._get(user_id)
+    def update(self, user_id: int, user_data: UserRepositoryUpdate) -> User:
+        user = self.__get(user_id)
         for field, value in user_data:
             setattr(user, field, value)
         self.session.commit()
         return user    
 
     def delete(self, user_id: int):
-        user = self._get(user_id)
+        user = self.__get(user_id)
+        if self.__has_book(user.id):
+            return Response(
+                status_code=status.HTTP_403_FORBIDDEN, 
+                content='{"Reason": "You can not delete user if he has a book"}', 
+                media_type='application/json')
         self.session.delete(user)
         self.session.commit()
         return Response(status_code=status.HTTP_204_NO_CONTENT)
